refactor(checker): split flow traversal
