+++ scripts/arch/arch_guard_shared.py
@@ -238,11 +238,7 @@ LINE_LIMIT_CHECKS = [
             # delete it from this set in the same diff and the
             # `test_excluded_files_actually_exceed_limit` test will catch
             # any regression.
-            "crates/tsz-checker/src/flow/control_flow/core.rs",
             "crates/tsz-checker/src/state/type_resolution/module.rs",
-            "crates/tsz-checker/src/types/class_type/constructor.rs",
-            "crates/tsz-checker/src/types/property_access_type/resolve.rs",
-            "crates/tsz-checker/src/types/type_checking/duplicate_identifiers.rs",
         },
     ),
     (
@@ -667,7 +663,7 @@ FILE_LINE_LIMIT_CHECKS = [
         / "flow"
         / "control_flow"
         / "core.rs",
-        2886,
+        1823,
     ),
     (
         "Solver boundary: type_queries/flow.rs size ratchet",
